consolidating shared code
